+++ apps/block_scout_web/assets/js/lib/add_chain_to_mm.js
@@ -15,8 +15,9 @@ export async function addChainToMM ({ btn }) {
     const jsonRPCObj = document.getElementById('js-json-rpc')
     // @ts-ignore
     const jsonRPC = jsonRPCObj && jsonRPCObj.value
+    const pathObj = document.getElementById('network-path')
     // @ts-ignore
-    const path = process.env.NETWORK_PATH || '/'
+    const path = (pathObj && pathObj.value) || '/'
 
     const blockscoutURL = location.protocol + '//' + location.host + path
     if (chainIDFromWallet !== chainIDFromInstance) {