+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 
 ### Chore
 
+- [#6954](https://github.com/blockscout/blockscout/pull/6954) - Move some compile time vars to runtime
 - [#6952](https://github.com/blockscout/blockscout/pull/6952) - Manage BlockReward fetcher params
 - [#6929](https://github.com/blockscout/blockscout/pull/6929) - Extend `INDEXER_MEMORY_LIMIT` env parsing
 - [#6902](https://github.com/blockscout/blockscout/pull/6902) - Increase verification timeout to 120 seconds for microservice verification